Incoherent thread synchronisation in file reporter
The file reporter is single threaded throught the disruptor EventHandler
There is no need for a thread local buffer cache.

The synchronized block is still there :-(
Someone with a good understanding of the Gravitee components lifecycle
should have a look at it.
If it's removed and the file reporter is shutdown before the reporting
disruptor then some writes may fail.